% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkout.R
\name{checkout}
\alias{checkout}
\title{Check a summarised_result object}
\usage{
checkout(result, output = "console")
}
\arguments{
\item{result}{A `summarised_result` object.}

\item{output}{Character vector, it can be:
\itemize{
  \item \code{"console"} to print the summary in the console.
  \item \code{"text"} to return a character vector.
  \item \code{"show"} to show the result in the viewer (html format).
  \item Otherwise \code{output} will be used as the name of an \code{'.md'}
    file to write the report in.
}}
}
\value{
Whether the result object passes the checks or not.
}
\description{
Check a summarised_result object
}
