% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{checkoutShiny}
\alias{checkoutShiny}
\title{Launch the OmopCheckout Shiny application}
\usage{
checkoutShiny()
}
\value{
Called for its side effect of launching the Shiny app. Returns
  \code{NULL} invisibly.
}
\description{
Opens an interactive Shiny application for uploading, exploring, and
validating `summarised_result` objects. The app provides the same checks
as [checkout()] in a point-and-click interface, without needing to write
R code. Also you can explore the results in the explore Tab.
}
