% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkout.R
\name{summaryEstimates}
\alias{summaryEstimates}
\title{Summarise results by estimate/strata}
\usage{
summaryEstimates(result, output = "console")
}
\arguments{
\item{result}{A `summarised_result` object.}

\item{output}{Either "console" to print the summary in the console, or "text"
to return a character vector or the name of an '.md' file to write the report
in.}
}
\value{
A character vector describing estimate/strata groupings and result ids.
}
\description{
Summarise results by estimate/strata
}
