% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkout.R
\name{summaryResult}
\alias{summaryResult}
\title{Summarise results by estimate/strata}
\usage{
summaryResult(result)
}
\arguments{
\item{result}{A `summarised_result` object.}
}
\value{
A character vector describing estimate/strata groupings and result ids.
}
\description{
Summarise results by estimate/strata
}
