% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkout.R
\name{checkout}
\alias{checkout}
\title{An}
\usage{
checkout(result, output = "console")
}
\arguments{
\item{result}{A `summarised_result` object.}

\item{output}{Either "console" to print the summary in the console, or "text"
to return a character vector or the name of an '.md' file to write the report
in.}
}
\value{
Whether the result object passes the checks or not.
}
\description{
An
}
