% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkout.R
\name{checkout}
\alias{checkout}
\title{An}
\usage{
checkout(result, file = NULL)
}
\arguments{
\item{result}{A `summarised_result` object.}

\item{file}{Name of an '.md' file to write the report.}
}
\value{
Whether the result object passes the checks or not.
}
\description{
An
}
