% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OmopCheckout-package.R
\docType{package}
\name{OmopCheckout-package}
\alias{OmopCheckout}
\alias{OmopCheckout-package}
\title{OmopCheckout: What the Package Does (Title Case)}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://OHDSI.github.io/OmopCheckout/}
  \item Report bugs at \url{https://github.com/OHDSI/OmopCheckout/issues}
}

}
\keyword{internal}
