% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkout.R
\name{summaryEstimates}
\alias{summaryEstimates}
\title{Summarise results by estimate/strata}
\usage{
summaryEstimates(result, output = "console")
}
\arguments{
\item{result}{A `summarised_result` object.}

\item{output}{Character vector, it can be:
\itemize{
  \item \code{"console"} to print the summary in the console.
  \item \code{"text"} to return a character vector.
  \item \code{"show"} to show the result in the viewer (html format).
  \item Otherwise \code{output} will be used as the name of an \code{'.md'}
    file to write the report in.
}}
}
\value{
A character vector describing estimate/strata groupings and result ids.
}
\description{
Summarise results by estimate/strata
}
