% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkout.R
\name{summaryPackages}
\alias{summaryPackages}
\title{Summarise package usage in a summarised_result object}
\usage{
summaryPackages(result, output = "console")
}
\arguments{
\item{result}{A `summarised_result` object.}

\item{output}{A single character string specifying where to send the output.
  It can be:
\itemize{
  \item \code{"console"} to print the summary in the console.
  \item \code{"text"} to return the summary as a character string.
  \item \code{"show"} to display the result in the viewer (HTML format).
  \item Otherwise \code{output} will be used as the name of an \code{'.md'}
    file to write the report to.
}}
}
\value{
A character string containing the package summary, returned
  invisibly unless \code{output = "text"}.
}
\description{
For each package name and version found in the result settings, reports the
number of rows and result IDs associated with it, broken down by result type.
}
