% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkout.R
\name{summarySuppress}
\alias{summarySuppress}
\title{Check suppression correctness for each result_id}
\usage{
summarySuppress(result, output = "console")
}
\arguments{
\item{result}{A `summarised_result` object.}

\item{output}{Character vector, it can be:
\itemize{
  \item \code{"console"} to print the summary in the console.
  \item \code{"text"} to return a character vector.
  \item \code{"show"} to show the result in the viewer (html format).
  \item Otherwise \code{output} will be used as the name of an \code{'.md'}
    file to write the report in.
}}
}
\value{
A character vector of messages (one per checked result id).
}
\description{
For each unique `result_id` in `result`, checks whether suppression is applied
and produces messages describing whether suppression is correct.
}
