% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkout.R
\name{summarySuppress}
\alias{summarySuppress}
\title{Check suppression correctness for each result_id}
\usage{
summarySuppress(result, output = "console")
}
\arguments{
\item{result}{A `summarised_result` object.}

\item{output}{Either "console" to print the summary in the console, or "text"
to return a character vector or the name of an '.md' file to write the report
in.}
}
\value{
A character vector of messages (one per checked result id).
}
\description{
For each unique `result_id` in `result`, checks whether suppression is applied
and produces messages describing whether suppression is correct.
}
